fix: rename Recording > RecordingApp
